third_party/WebKit/LayoutTests/editing/spelling/paste_into_element_starting_with_invisible.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="resources/util.js"></script>

<div id="source">
zz.
</div>

<div contenteditable>
<span id="target">[]</span>
</div>

<script>
// Regression test for crbug.com/640456

function copyAndPaste() {
    var source = document.getElementById('source');
    var target = document.getElementById('target');
    var sel = window.getSelection();
    sel.selectAllChildren(source);
    document.execCommand('copy');
    sel.collapse(target.firstChild, 1);
    document.execCommand('paste');
}

runSpellingTest(
    [() => copyAndPaste()],
    [() => assert_true(internals.hasSpellingMarker(document, 1, 2))],
    'Spelling markers should appear immediately after pasting.');
</script>
